backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
feedparser==6.0.11
apscheduler==3.10.4
mistralai==1.0.3
python-dotenv==1.0.1
cachetools==5.3.3
praw==7.7.1
